WebIDL/tests/submissions/heycam/extensible-001.html

<!DOCTYPE html>
<title>Objects are extensible</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4. ECMAScript binding
//
// Objects are [[Extensible]] = true, by default.  And there aren't any
// objects that are defined to be non-extensible.

var objects = [

  // interface object

  // feature: DOMImplementation (interface) (without inheritance, without [NoInterfaceObject])
  interfaceObject("DOMImplementation"), "DOMImplementation interface object",
  // feature: HTMLCollection (interface) (without inheritance, without [NoInterfaceObject])
  interfaceObject("HTMLCollection"), "HTMLCollection interface object",


  // interface object for inherited interface

  // feature: Element (interface) (with inheritance, without [NoInterfaceObject])
  interfaceObject("Element"), "Element interface object",
  // feature: UIEvent (interface) (with inheritance, without [NoInterfaceObject])
  interfaceObject("UIEvent"), "UIEvent interface object",


  // interface prototype object

  // feature: DOMImplementation (interface) (without inheritance)
  interfacePrototypeObject("DOMImplementation"), "DOMImplementation interface prototype object",
  // feature: HTMLCollection (interface) (without inheritance)
  interfacePrototypeObject("HTMLCollection"), "HTMLCollection interface prototype object",


  // interface prototype object for inherited interface

  // feature: Element (interface) (with inheritance)
  interfacePrototypeObject("Element"), "Element interface prototype object",
  // feature: UIEvent (interface) (with inheritance)
  interfacePrototypeObject("UIEvent"), "UIEvent interface prototype object",


  // named constructor object

  // feature: HTMLImageElement (interface) (with [NamedConstructor])
  namedConstructorObject("Image"), "Image named constructor object",
  // feature: HTMLAudioElement (interface) (with [NamedConstructor])
  namedConstructorObject("Audio"), "Audio named constructor object",


  // attribute getter

  // feature: Document.title (attribute) (writable)
  getter("Document", "title"), "Document.title attribute getter",
  // feature: HTMLSelectElement.length (attribute) (writable)
  getter("HTMLSelectElement", "length"), "HTMLSelectElement.length attribute getter",


  // read only attribute getter

  // feature: Node.parentNode (attribute) (read only)
  getter("Node", "parentNode"), "Node.parentNode attribute getter",
  // feature: HTMLFormElement.length (attribute) (read only)
  getter("HTMLFormElement", "length"), "HTMLFormElement.length attribute getter",


  // attribute setter

  // feature: Document.title (attribute) (writable)
  setter("Document", "title"), "Document.title attribute setter",
  // feature: HTMLSelectElement.length (attribute) (writable)
  setter("HTMLSelectElement", "length"), "HTMLSelectElement.length attribute setter",


  // operation function

  // feature: HTMLElement.click (operation)
  operation("HTMLElement", "click"), "HTMLElement.click function",
  // feature: Node.appendChild (operation)
  operation("Node", "appendChild"), "Node.appendChild function",


  // stringifier function

  // feature: DOMTokenList (stringifier)
  stringifier("DOMTokenList"), "DOMTokenList stringifier function",


  // exception interface object

  // feature: DOMException (exception) (without [NoInterfaceObject])
  interfaceObject("DOMException"), "DOMException exception interface object",


  // exception interface prototype object

  // feature: DOMException (exception)
  interfacePrototypeObject("DOMException"), "DOMException exception interface prototype object",


  // exception field getter

  // feature: DOMException.code (exception field)
  getter("DOMException", "code"), "DOMException.code exception field getter",


  // platform object implementing an interface

  // feature: DOMImplementation (interface) (without inheritance, primary interface)
  function() { return document.implementation; }, "a platform object implementing DOMImplementation",
  // feature: HTMLCollection (interface) (without inheritance, primary interface)
  function() { return document.getElementsByTagName("div"); }, "a platform object implementing HTMLCollection",


  // platform object implementing an inherited interface

  // feature: HTMLDivElement (interface) (with inheritance, primary interface)
  function() { return document.createElement("div"); }, "a platform object implementing Element",
  // feature: UIEvent (interface) (with inheritance, primary interface)
  function() { return document.createEvent("UIEvent"); }, "a platform object implementing UIEvent",


  // platform object representing an exception

  // feature: DOMException (exception)
  function() {
    try {
      document.appendChild(document);
    } catch (e) {
      return e;
    }
  }, "a platform object representing a DOMException",


  // platform array object
  // XXX


  // named properties object

  // feature: Window (interface) (with [NamedPropertiesObject])
  function() { return Object.getPrototypeOf(Window.prototype); }, "named properties object"
];

for (var i = 0; i < objects.length; i += 2) {
  test(function() {
    var object = objects[i]();
    assert_true(Object.isExtensible(object), "object is extensible");
  }, objects[i + 1] + " is extensible");
}
</script>
